Editar notas, en contruccion

+++ app/Http/Controllers/MateriaController.php
@@ -97,22 +97,24 @@ class MateriaController extends Controller
      */
     public function destroy($id)
     {
-        if(session('role')=="administrador"){
-            // Solo el administrador puede eliminar una MateriaPC
-            $materia = Materia::where("pk_materia","=",$id)->get();
-            if(empty($materia[0])){
+        if($request->ajax()){
+            if(session('role')=="administrador"){
+                // Solo el administrador puede eliminar una MateriaPC
+                $materia = Materia::where("pk_materia","=",$id)->get();
+                if(empty($materia[0])){
+                    return response()->json([
+                        'mensaje' => 'Esta materia no existe.'
+                    ]);
+                }
+                $materia[0]->delete();
                 return response()->json([
-                    'mensaje' => 'Esta materia no existe.'
+                    'mensaje' => 'La materia fue eliminada.'
                 ]);
+                
             }
-            $materia[0]->delete();
             return response()->json([
-                'mensaje' => 'La materia fue eliminada.'
+                'mensaje' => 'No tienes los permisos necesarios.'
             ]);
-            
         }
-        return response()->json([
-            'mensaje' => 'No tienes los permisos necesarios.'
-        ]);
     }
 }
